MeMoPlayer: Secure RecordStores now start with an @ instead of an _
This was causing the default RecordStore (named _EMPTY) to be a secure
RecordStore which slowdowns the common usage.

This patch also fixes an error where the erase() method would still be
called on the secure RecordStore (which we don't want) when RecordStore
was in the list of the opened RecordStores "m_instance".