Improved automatic image purge in Scene.
Only purge images according to their compressed size (from m_data).
Only take in account the real size of the image when NOT using WeakReference.

This lets the GC free the Image objects automatically instead of using
the purge mechanism that will remove all image data (the compressed PNG)
and will force the scene to re-download the image again.