Fix broken Style.getColor() by 1642df (Improved SF/MFColor implementation)
Using an integer type register to handle SFColor fields broke the
Style.getColor() method that only works on receiving a Field type register.
So for compatibility, we will keep using the Field type register on SFColor.

Comparing a SFColor to another SFColor or an MFColor index is therefore
not possible anymore (without hacking the Register class).