Keep reference to cache manager even after deletion for RMS2
As Namespace keep a weak reference to the cache manager (even when
deleted), we cannot remove the cache manager from the list.
This prevents opening the RMS multiple times if a new RMSCacheManager2
instance is created while an older one is kept in the weak ref by the
Namespace. This can happen on widget deletion then reinstall.