Fix message propagation bug on Nokia N70
Bug: On first application launch, Nokia N70 will have all calls to
hashCode() return the same value !

Fix: Don't use hashCode() to check for object uniqueness, use a custom
integer instead.

We need to track which Message object triggered the event, but we can
only keep an integer in ObjLink. We previously used hashCode() to get
a unique integer id per object, but this is not supported on some
platforms like the Nokia N70 which return always the same value(!!).