Render BitmapFont with drawImage when drawRGB does not support blending
Each glyph will keep a ref to a cached Image and rendered using drawImage.
This is less CPU expensive than the previous approach (see ExternFontPatch
that kept the RGB data for each glyphs) and keeps the nominal case
untouched (render a subregion of the big RGB array with drawRGB).

This feature is automatically enabled on phones that claim to not support
blending as some of them still support it with drawImage.
This feature can enabled manually with the "MeMo-BitmapFont"
property set to "image" (any other value will disable it.